improve code for color convert.

+++ color.h
@@ -9,8 +9,8 @@ extern "C" {
 #endif
 
 void init_color_convert(void);
-void yuv_to_rgb(BYTE y, BYTE u, BYTE v, BYTE *r, BYTE *g, BYTE *b);
-void rgb_to_yuv(BYTE r, BYTE g, BYTE b, BYTE *y, BYTE *u, BYTE *v);
+void yuv_to_rgb(int y, int u, int v, BYTE *r, BYTE *g, BYTE *b);
+void rgb_to_yuv(BYTE r, BYTE g, BYTE b, int *y, int *u, int *v);
 
 #ifdef __cplusplus
 }
